Refactor: 특정함수에 맞게 file 변경 및 html statement 추가
