+++ ex_1/createStatementData.js
@@ -1,0 +1,61 @@
+function createStatementData(invoice, plays) {
+    const statementData = {};
+
+    statementData.customer = invoice.customer;
+    statementData.performances = invoice.performances.map(enrichPerformance);
+    statementData.totalAmount = totalAmount(statementData);
+    statementData.totalVolumeCredits = totalVolumeCredits(statementData);
+    return statementData; // 중간 데이터 구조를 인수로 전달
+    function enrichPerformance(aPerformance) {
+        const result = Object.assign({}, aPerformance); //얕은 복사 수행
+        //첫 레벨만 복사해서 가져옴.
+        result.play = playFor(result);
+        result.amount = amountFor(result);
+        result.volumeCredits = volumeCreditsFor(result);
+        return result;
+    }
+
+    //플레이어 추출함수.
+    function playFor(aPerformance) {
+        return plays[aPerformance.playID];
+    }
+
+    function amountFor(aPerformance) {
+        let result = 0;
+
+        switch (aPerformance.play.type) {
+            case 'tragedy':
+                result = 40000;
+                if (aPerformance.audience > 30) {
+                    result += 1000 * (aPerformance.audience - 30);
+                }
+                break;
+            case 'comedy':
+                result = 30000;
+                if (aPerformance.audience > 20) {
+                    result += 10000 + 500 * (aPerformance.audience - 20);
+                }
+                result += 300 * aPerformance.audience;
+                break;
+            default:
+                throw new Error(`unknown type: ${aPerformance.play.type}`);
+        }
+        return result; //함수 안에서 값이 변경되는 변수 반환.
+    }
+    //volume creadits 구하는 함수
+    function volumeCreditsFor(aPerformance) {
+        let result = 0;
+        result += Math.max(aPerformance.audience - 30, 0);
+        if ('comedy' === aPerformance.play.type) result += Math.floor(perf.audience / 5);
+
+        return result;
+    }
+
+    function totalAmount(data) {
+        return data.performances.reduce((total, p) => total + p.amount, 0);
+    }
+
+    function totalVolumeCredits(data) {
+        return data.performances.reduce((total, p) => total + p.volumeCredits, 0);
+    }
+}
